requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
PySocks==1.5.0
Werkzeug==0.9.4
cssselect==0.9.1
gunicorn==18.0
itsdangerous==0.23
lxml==3.3.3
psycopg2==2.5.2
pyquery==1.2.8
selenium==2.40.0
